Move CommandActionMapper to cardslist package.
