+++ remote-messaging/remote-messaging-impl/src/test/java/com/duckduckgo/remote/messaging/impl/modal/cardslist/RealCommandActionMapperTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.remote.messaging.impl.modal
+package com.duckduckgo.remote.messaging.impl.modal.cardslist
 
 import com.duckduckgo.remote.messaging.api.Action
 import com.duckduckgo.remote.messaging.impl.modal.cardslist.CardsListRemoteMessageViewModel.Command
